Allow a custom prefix in the issue title.
One cool thing about exception notification is the option to configure an email prefix. So in the case where we've got several servers, we can add a different prefix for each different server, and immediately know where the exception happened when we receive the email.

In some cases this option is better than using extra labels (or complements it), especially when users rely on GitHub's email
notifications.

This adds a similar functionality to Party Foul.